Split Jinrai and Dragonlash into two moves
