+++ fighters/ken/src/status.rs
@@ -7,6 +7,8 @@ mod attack;
 mod attack_lw4_start;
 mod attack_lw4;
 
+mod attack_command1;
+
 mod special_n2;
 
 mod special_s;
@@ -24,6 +26,8 @@ pub fn install(agent: &mut Agent) {
 
     attack_lw4_start::install(agent);
     attack_lw4::install(agent);
+
+    attack_command1::install(agent);
 
     special_n2::install(agent);
